requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
aiogram==3.13.1
motor==3.3.1
pymongo==4.8.0
python-dotenv==1.0.1
pytest==7.4.4
pytest-asyncio==0.23.8
